Move `processArguments()` to process-arguments.js
